Configured rtk query & API's

+++ frontend/src/utilities/Datefunc.ts
@@ -1,7 +1,7 @@
 export function formatDate(date: string): string {
     const formattedDate = new Date(date);
     return formattedDate.toLocaleString("en-GB", {
-        weekday: "long",
+        weekday: "short",
         year: "numeric",
         month: "short",
         day: "numeric",
